requirements.txt
# Database and PostgreSQL connectivity
psycopg2-binary==2.9.6

# Kafka integration
kafka-python==2.0.2

# Spark (for processing streaming data)
pyspark==3.3.0

# HTTP client for API calls
requests==2.31.0

# Data manipulation and analysis
pandas==2.0.3

# Visualization libraries
matplotlib==3.6.3
seaborn==0.12.2

# SQLAlchemy for database connections in visualization
sqlalchemy==2.0.18

# Code quality and formatting tools (optional, for development)
black==23.3.0
flake8==6.0.0
pylint==2.17.3
